+++ Sources/Swift/Transactions/SentrySpanKey.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@objcMembers @objc(SentrySpanKey)
+class SentrySpanKey: NSObject {
+    /// Used to set the number of bytes processed in a file span operation
+    static let fileSize = "file.size"
+}