+++ Tests/SentryTests/Transactions/SentrySpanKeyTests.swift
@@ -1,0 +1,8 @@
+@testable import Sentry
+import XCTest
+
+class SentrySpanKeyTests: XCTestCase {
+    func testAppLifecycle_shouldBeExpectedValue() {
+        XCTAssertEqual(SentrySpanKey.fileSize, "file.size")
+    }
+}